+++ services/QuillLMS/app/services/learn_worlds_integration/user_request.rb
@@ -1,0 +1,24 @@
+# frozen_string_literal: true
+
+module LearnWorlds
+  class UserRequest < ::ApplicationService
+    class NilUserError < StandardError; end
+
+    attr_reader :user
+
+    def initialize(user)
+      @user = user
+    end
+
+    def run
+      raise NotImplementedError
+    end
+
+    private def headers
+      {
+        "Lw-Client" => CLIENT_ID,
+        "Authorization" => "Bearer #{ACCESS_TOKEN}"
+      }
+    end
+  end
+end
